extract login from controller into service
