+++ admin/controllers/authentication.js
@@ -4,10 +4,7 @@ const logger = require('../services/log.service').getLogger()
 const homeRoutes = require('../lib/consts/home-routes')
 const authModes = require('../lib/consts/auth-modes')
 const config = require('../config')
-const bluebird = require('bluebird')
-const jwt = bluebird.promisifyAll(require('jsonwebtoken'))
-const request = require('async-request')
-const roleService = require('../services/role.service')
+const dfeSignInService = require('../services/dfe-signin.service')
 const url = require('url')
 const passport = require('passport')
 
@@ -59,53 +56,13 @@ const getSignIn = (req, res) => {
   }
 }
 
-const createJwtForDfeApi = async () => {
-  const clientId = config.Auth.dfeSignIn.clientId
-  const apiSecret = config.Auth.dfeSignIn.userInfoApi.apiSecret
-  const payload = {
-    iss: clientId,
-    aud: config.Auth.dfeSignIn.userInfoApi.audience
-  }
-  return jwt.sign(payload, apiSecret, { algorithm: 'HS256' })
-}
-
-const getUserInfoFromDfeApi = async (token, user) => {
-  const serviceId = config.Auth.dfeSignIn.clientId // serves as serviceId also, undocumented
-  const orgId = user.organisation.id
-  const baseUrl = config.Auth.dfeSignIn.userInfoApi.baseUrl
-  const url = `${baseUrl}/services/${serviceId}/organisations/${orgId}/users/${user.id}`
-  const response = await request(url, {
-    method: 'GET',
-    headers: {
-      Authorization: `Bearer ${token}`
-    }
-  })
-  if (response.statusCode === 200) {
-    return JSON.parse(response.body)
-  } else {
-    logger.error(response)
-    throw new Error(`unsatisfactory response returned from DfE API. statusCode:${response.statusCode}`)
-  }
-}
-
 const postDfeSignIn = async (req, res) => {
-  logger.debug('## req.user ##')
-  logger.debug(JSON.stringify(req.user, null, 2))
-  try {
-    // get role info...
-    const token = await createJwtForDfeApi()
-    const userInfo = await getUserInfoFromDfeApi(token, req.user)
-    logger.debug('## userInfo from API ##')
-    logger.debug(JSON.stringify(userInfo, null, 2))
-    // TODO array check
-    const mtcRole = roleService.mapDfeRoleToMtcRole(userInfo.roles[0].code)
-    req.user.role = mtcRole
-    logger.debug(`user role is ${req.user.role}`)
-    logger.info(`postSignIn: User ID logged in: ${req.user.displayName} (${req.user.id})
-    timezone is "${req.user.timezone}"`)
-  } catch (error) {
-    throw new Error(`unable to resolve dfe user:${error.message}`)
-  }
+  dfeSignInService.process(req.user)
+  logger.info(`postSignIn: User ID logged in:
+    id:${req.user.id} \n
+    displayName:${req.user.displayName} \n
+    role:${req.user.role} \n
+    timezone:"${req.user.timezone}"`)
 
   switch (req.user.role) {
     case 'TEACHER':
@@ -122,10 +79,13 @@ const postDfeSignIn = async (req, res) => {
 }
 
 const postSignIn = (req, res) => {
-  const { displayName, id, role, timezone } = req.user
-  logger.info(`postSignIn: User ID logged in: ${id} (${displayName}) timezone is "${timezone}"`)
+  logger.info(`postSignIn: User ID logged in:
+  id:${req.user.id} \n
+  displayName:${req.user.displayName} \n
+  role:${req.user.role} \n
+  timezone:"${req.user.timezone}"`)
 
-  switch (role) {
+  switch (req.user.role) {
     case 'TEACHER':
     case 'HEADTEACHER':
     case 'HELPDESK':
@@ -141,6 +101,9 @@ const postSignIn = (req, res) => {
 
 const getSignOut = (req, res) => {
   req.logout()
+  if (config.Auth.Mode === authModes.dfeSignIn) {
+    getDfeSignOut(req, res)
+  }
   req.session.regenerate(function () {
     // session has been regenerated
     switch (config.Auth.mode) {
@@ -148,7 +111,8 @@ const getSignOut = (req, res) => {
         res.redirect(config.Auth.ncaTools.authUrl)
         break
       case authModes.dfeSignIn:
-        getDfeSignOut(req, res)
+        // 🤷‍♂️
+        logger.debug('session regen called')
         break
       default: //  local
         res.render('/')
@@ -159,12 +123,6 @@ const getSignOut = (req, res) => {
 
 const getDfeSignOut = (req, res) => {
   if (req.user && req.user.id_token) {
-    logger.audit('User logged out', {
-      type: 'Sign-out',
-      userId: req.user.sub,
-      email: req.user.email,
-      client: 'profiles'
-    })
     const idToken = req.user.id_token
     const issuer = passport._strategies.oidc._issuer
     // let returnUrl = `${config.hostingEnvironment.protocol}://${config.hostingEnvironment.host}:${config.hostingEnvironment.port}/signout/complete`
